Update version to 0.1.126 and fix critical bug in neuron impact calculation
- Bump version in Cargo.toml from 0.1.125 to 0.1.126.
- Implement a critical bug fix in the neuron impact calculation, correcting the formula to use absolute weights instead of normalized weights. This change prevents severe underestimation of neuron contributions, which previously led to incorrect removal predictions.
- Update README to document the bug fix, including detailed explanations and examples of the impact calculation changes.
- Remove the build_inbound_weights function as it is no longer needed with the new calculation method.
- Adjust tests to reflect the updated impact calculation logic.

These changes enhance the accuracy of neuron impact assessments and improve the overall reliability of the analysis framework.
