+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "neat_ai_discovery"
-version = "0.1.125"
+version = "0.1.126"
 edition = "2021"
 
 [lib]
